pyproject.toml
[project]
name = "aragorn1025-tw-stock-crawler"
version = "0.1.0"
description = ""
authors = [
    {name = "aragorn1025",email = "tzjie.dai@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client (>=2.179.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
]

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
pylint = "^3.3.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py313"]
skip-string-normalization = false
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
  | \.mypy_cache
  | \.pytest_cache
  | .*/migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 120
combine_as_imports = true
skip = [
  ".venv",
  "build",
  "dist",
  "__pycache__",
  ".mypy_cache",
  ".pytest_cache",
  "migrations",
]

[tool.pylint.main]
generated-members = [
  "googleapiclient.discovery.*",
  "pandas.*", 
]

[tool.pylint.design]
min-public-methods = 1

[tool.pylint.format]
max-line-length = 120
